lab: Remove git version from version command
The 'lab version' command prints out the git version.  For example,

[prarit@prarit ~]$ lab version
git version 2.29.2
lab version master

Remove the git version output and move the option into the root file.
